+++ crates/api_models/src/payments.rs
@@ -184,6 +184,7 @@ pub struct PaymentsRequest {
     pub metadata: Option<Metadata>,
 
     /// Information about the product , quantity and amount for connectors. (e.g. Klarna)
+    #[schema(value_type = Option<Vec<OrderDetailsWithAmount>>, example = "order_details as vector of objects")]
     pub order_details: Option<Vec<OrderDetailsWithAmount>>,
 
     /// It's a token used for client side verification.
@@ -1242,6 +1243,7 @@ pub struct PaymentsResponse {
     pub metadata: Option<pii::SecretSerdeValue>,
 
     /// Information about the product , quantity and amount for connectors. (e.g. Klarna)
+    #[schema(value_type = Option<Vec<OrderDetailsWithAmount>>, example = "order_details as vector of objects")]
     pub order_details: Option<Vec<pii::SecretSerdeValue>>,
 
     /// description: The customer's email address
