+++ crates/router/src/connector/adyen/transformers.rs
@@ -900,11 +900,11 @@ fn get_line_items(item: &types::PaymentsAuthorizeRouterData) -> Vec<LineItem> {
     match order_details {
         Some(od) => od
             .iter()
-            .map(|data| LineItem {
-                amount_including_tax: Some(item.request.amount),
-                amount_excluding_tax: Some(item.request.amount),
+            .map(|data: &payments::OrderDetailsWithAmount| LineItem {
+                amount_including_tax: Some(data.amount),
+                amount_excluding_tax: Some(data.amount),
                 description: Some(data.product_name.clone()),
-                id: Some(String::from("Items #1")),
+                id: Some(common_utils::generate_id(consts::ID_LENGTH, "item_")),
                 tax_amount: None,
                 quantity: Some(data.quantity),
             })
@@ -914,7 +914,7 @@ fn get_line_items(item: &types::PaymentsAuthorizeRouterData) -> Vec<LineItem> {
                 amount_including_tax: Some(item.request.amount),
                 amount_excluding_tax: Some(item.request.amount),
                 description: None,
-                id: Some(String::from("Items #1")),
+                id: Some(common_utils::generate_id(consts::ID_LENGTH, "item_")),
                 tax_amount: None,
                 quantity: Some(1),
             };
